+++ R/backend-redshift.R
@@ -4,6 +4,12 @@ sql_translate_env.RedshiftConnection <- function(con) {
 
   sql_variant(
     sql_translator(.parent = postgres$scalar,
+
+      # https://stackoverflow.com/questions/56708136
+      paste  = sql_paste_redshift(" "),
+      paste0 = sql_paste_redshift(""),
+      str_c = sql_paste_redshift(""),
+
       # https://docs.aws.amazon.com/redshift/latest/dg/REGEXP_REPLACE.html
       str_replace = sql_not_supported("str_replace"),
       str_replace_all = function(string, pattern, replacement) {
@@ -17,3 +23,7 @@ sql_translate_env.RedshiftConnection <- function(con) {
 
 #' @export
 sql_translate_env.Redshift <- sql_translate_env.RedshiftConnection
+
+sql_paste_redshift <- function(sep) {
+  sql_paste_infix(sep, "||", function(x) sql_expr(cast(!!x %as% text)))
+}