+++ tests/testthat/test-backend-redshift.R
@@ -8,4 +8,8 @@ test_that("string translations", {
 
   expect_error(translate_sql(str_replace("xx", ".", "a")), "not available")
   expect_equal(translate_sql(str_replace_all("xx", ".", "a")), sql("REGEXP_REPLACE('xx', '.', 'a')"))
+
+  expect_equal(translate_sql(paste("x", "y")), sql("'x' || ' ' || 'y'"))
+  expect_equal(translate_sql(paste0("x", "y")), sql("'x' || 'y'"))
+  expect_equal(translate_sql(str_c("x", "y")), sql("'x' || 'y'"))
 })